refactor: order middleware to separate file
